refactor(be): profileRepository 경로 오타 수정
