+++ backend/src/main/java/com/deliveranything/domain/user/profile/repository/CustomerAddressRepository.java
@@ -1,4 +1,4 @@
-package com.deliveranything.domain.user.profile.rerpository;
+package com.deliveranything.domain.user.profile.repository;
 
 import com.deliveranything.domain.user.profile.entity.CustomerAddress;
 import com.deliveranything.domain.user.profile.entity.CustomerProfile;
