+++ backend/src/main/java/com/deliveranything/domain/user/profile/repository/RiderProfileRepository.java
@@ -1,4 +1,4 @@
-package com.deliveranything.domain.user.profile.rerpository;
+package com.deliveranything.domain.user.profile.repository;
 
 import com.deliveranything.domain.user.profile.entity.RiderProfile;
 import java.util.Optional;
