+++ backend/src/main/java/com/deliveranything/domain/user/profile/service/CustomerProfileService.java
@@ -4,9 +4,9 @@ import com.deliveranything.domain.user.profile.entity.CustomerAddress;
 import com.deliveranything.domain.user.profile.entity.CustomerProfile;
 import com.deliveranything.domain.user.profile.entity.Profile;
 import com.deliveranything.domain.user.profile.enums.ProfileType;
-import com.deliveranything.domain.user.profile.rerpository.CustomerAddressRepository;
-import com.deliveranything.domain.user.profile.rerpository.CustomerProfileRepository;
-import com.deliveranything.domain.user.profile.rerpository.ProfileRepository;
+import com.deliveranything.domain.user.profile.repository.CustomerAddressRepository;
+import com.deliveranything.domain.user.profile.repository.CustomerProfileRepository;
+import com.deliveranything.domain.user.profile.repository.ProfileRepository;
 import com.deliveranything.domain.user.user.entity.User;
 import com.deliveranything.domain.user.user.repository.UserRepository;
 import java.util.List;
