+++ backend/src/main/java/com/deliveranything/domain/user/profile/service/RiderProfileService.java
@@ -4,8 +4,8 @@ import com.deliveranything.domain.user.profile.entity.Profile;
 import com.deliveranything.domain.user.profile.entity.RiderProfile;
 import com.deliveranything.domain.user.profile.enums.ProfileType;
 import com.deliveranything.domain.user.profile.enums.RiderToggleStatus;
-import com.deliveranything.domain.user.profile.rerpository.ProfileRepository;
-import com.deliveranything.domain.user.profile.rerpository.RiderProfileRepository;
+import com.deliveranything.domain.user.profile.repository.ProfileRepository;
+import com.deliveranything.domain.user.profile.repository.RiderProfileRepository;
 import com.deliveranything.domain.user.user.entity.User;
 import com.deliveranything.domain.user.user.repository.UserRepository;
 import com.deliveranything.global.exception.CustomException;
