+++ backend/src/main/java/com/deliveranything/domain/user/profile/service/SellerProfileService.java
@@ -3,8 +3,8 @@ package com.deliveranything.domain.user.profile.service;
 import com.deliveranything.domain.user.profile.entity.Profile;
 import com.deliveranything.domain.user.profile.entity.SellerProfile;
 import com.deliveranything.domain.user.profile.enums.ProfileType;
-import com.deliveranything.domain.user.profile.rerpository.ProfileRepository;
-import com.deliveranything.domain.user.profile.rerpository.SellerProfileRepository;
+import com.deliveranything.domain.user.profile.repository.ProfileRepository;
+import com.deliveranything.domain.user.profile.repository.SellerProfileRepository;
 import com.deliveranything.domain.user.user.entity.User;
 import com.deliveranything.domain.user.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
